requirements.txt
absl-py==2.1.0
altgraph==0.17.4
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
filelock==3.13.1
fonttools==4.54.1
fsspec==2024.2.0
ftfy==6.2.3
grpcio==1.66.2
huggingface-hub==0.25.1
idna==3.10
imageio==2.35.1
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
opencv-python==4.10.0.84
packaging==24.1
pefile==2024.8.26
pillow==10.2.0
protobuf==5.28.2
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
setuptools==74.1.2
six==1.16.0
sympy==1.12
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tifffile==2024.9.20
tokenizers==0.19.1
torch==2.4.0+cu121
torchaudio==2.4.1+cu121
torchvision==0.19.1+cu121
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.9.0
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
wheel==0.44.0
